requirements.txt
langchain==0.1.20
langchain-google-genai==1.0.6
langchain-community==0.0.38
google-generativeai==0.7.2
chromadb==0.4.24
faiss-cpu==1.8.0
pandas==2.0.3
streamlit==1.32.2
python-dotenv==1.0.1
sentence-transformers==2.7.0
tiktoken==0.6.0